Workaround for Ubuntu dock window switch
When using Super + Number to switch to an application having several windows,
you use arrows to chose which window you want to focus, then you press Enter
to switch to that window. In case case, Enter key could be stuck repeating.

This is a workaround to press Escape after Enter, breaking the repeating Enter.

I'm sorry I have to resort to a dirty solution,
but it does the trick for me :D